LayoutTests/imported/w3c/web-platform-tests/navigation-api/focus-reset/basic-expected.txt


PASS Invalid values for focusReset throw
PASS Does not reset the focus when no navigate handler is present
FAIL Resets the focus when no focusReset option is provided assert_equals: Focus stays on the button during the transition expected Element node <button></button> but got Element node <body><button></button><button tabindex="0"></button></body>
FAIL Resets the focus when focusReset is explicitly set to undefined assert_equals: Focus stays on the button during the transition expected Element node <button></button> but got Element node <body><button></button><button tabindex="0"></button></body>
PASS Resets the focus when no focusReset option is provided (nontrivial fulfilled promise)
FAIL Resets the focus when no focusReset option is provided (rejected promise) assert_equals: Focus stays on the button during the transition expected Element node <button></button> but got Element node <body><button></button><button tabindex="0"></button></body>
FAIL Resets the focus when focusReset is explicitly set to 'after-transition' assert_equals: Focus stays on the button during the transition expected Element node <button></button> but got Element node <body><button></button><button tabindex="0"></button></body>
PASS Does not reset the focus when focusReset is set to 'manual'
